Refactor out most of the interface bits to main.rs
This moves all the interface bits to main.rs, leaving only the data
analysis logic in corresponding modules. Crayon is still badly broken.
